requirements.txt
beanie
discord
python-dotenv